Move to multi-chunk strategy
