+++ src/amphtml/amphtml.js
