--- src/index.js
@@ -1,5 +1,0 @@
-// `./Amp` is a codegen-generated file
-// eslint-disable-next-line import/export
-export * as amphtml from './amphtml';
-export * as helpers from './helpers';
-export * as setup from './setup';
